re_log
Module setup

Function to setup logging in binaries and web apps.

Structs

PanicOnWarn 🔒

PanicOnWarnScope

Scope for enabling panic on warn/error log messages temporariliy on the current thread (!).

Constants

LOG_FILE_LINE 🔒

PANIC_ON_WARN_SCOPE_DEPTH 🔒

Functions

env_var_bool 🔒

setup_logging

Sets up logging for the current process using default log filter as defined in crate::default_log_filter.

setup_logging_with_filter

Sets up logging for the current process using an explicit log filter.