Crate re_log

Text logging (nothing to do with rerun logging) for use in rerun libraries.

Provides helpers for adding multiple loggers, and for setting up logging on native and on web.

trace: spammy things

debug: things that might be useful when debugging

info: things that we want to show to users

warn: problems that we can recover from

error: problems that lead to loss of functionality or data

The warn_once etc macros are for when you want to suppress repeated logging of the exact same message.

Modules

channel_logger 🔒

Capture log messages and send them to some receiver over a channel.

external

Re-exports of other crates.

multi_logger 🔒

Have multiple loggers implementing log::Log at once.

result_extensions 🔒

setup 🔒

Function to setup logging in binaries and web apps.

Macros

debug

Constructs an event at the debug level.

debug_once

Logs a message once at the debug level.

error

Constructs an event at the error level.

error_once

Logs a message once at the error level.

info

Constructs an event at the info level.

info_once

Logs a message once at the info level.

trace

Constructs an event at the trace level.

trace_once

Logs a message once at the trace level.

warn

Constructs an event at the warn level.

warn_once

Logs a message once at the warn level.

Structs

ChannelLogger

Pipe log messages to a channel.

LogMsg

MultiLoggerNotSetupError

Produced when trying to install additional loggers when crate::setup_native_logging has not been called.

Enums

Level

An enum representing the available verbosity levels of the logger.

LevelFilter

An enum representing the available verbosity level filters of the logger.

Constants

CRATES_AT_ERROR_LEVEL 🔒

Never log anything less serious than a ERROR from these crates.

CRATES_AT_INFO_LEVEL 🔒

Never log anything less serious than a INFO from these crates.

CRATES_AT_WARN_LEVEL 🔒

Never log anything less serious than a WARN from these crates.

Traits

ResultExt

Functions

add_boxed_logger

Install an additional global logger.

add_logger

Install an additional global logger.

default_log_filter

Get RUST_LOG environment variable or info, if not set.

is_log_enabled 🔒

Should we log this message given the filter?

setup_native_logging

Directs log calls to stderr.

shorten_file_path 🔒

Shorten a path to a Rust source file.